finish splitting tests
