+++ test/index.js
@@ -17,53 +17,6 @@ describe('demo app', function () {
 
   let app
 
-  const setupApp = function (app) {
-    app.client.addCommand('dismissAboutPage', function () {
-      return this.isVisible('.js-nav').then(function (navVisible) {
-        if (!navVisible) {
-          return this.click('button[id="get-started"]').pause(500)
-        }
-      })
-    })
-
-    app.client.addCommand('selectSection', function (section) {
-      return this.click('button[data-section="' + section + '"]').pause(100)
-        .waitForVisible('#' + section + '-section')
-    })
-
-    app.client.addCommand('expandDemos', function () {
-      return this.execute(function () {
-        for (let demo of document.querySelectorAll('.demo-wrapper')) {
-          demo.classList.add('is-open')
-        }
-      })
-    })
-
-    app.client.addCommand('collapseDemos', function () {
-      return this.execute(function () {
-        for (let demo of document.querySelectorAll('.demo-wrapper')) {
-          demo.classList.remove('is-open')
-        }
-      })
-    })
-
-    app.client.addCommand('auditSectionAccessibility', function (section) {
-      const options = {
-        ignoreRules: ['AX_COLOR_01', 'AX_TITLE_01']
-      }
-      return this.selectSection(section)
-        .expandDemos()
-        .auditAccessibility(options).then(function (audit) {
-          if (audit.failed) {
-            throw Error(section + ' section failed accessibility audit\n' + audit.message)
-          }
-        })
-    })
-
-    chaiAsPromised.transferPromiseness = app.transferPromiseness
-    return app.client.waitUntilWindowLoaded()
-  }
-
   const startApp = () => {
     app = new Application({
       path: electron,
@@ -73,10 +26,16 @@ describe('demo app', function () {
       waitTimeout: timeout
     })
 
-    return app.start().then(setupApp)
+    return app.start().then((ret) => {
+      setup.setupApp(ret)
+    })
   }
 
-  const restartApp = () => app.restart().then(setupApp)
+  const restartApp = () => {
+    app.restart().then((ret) => {
+      setup.setupApp(ret)
+    })
+  }
 
   before(() => {
     setup.removeStoredPreferences()
